remove main.ts, refactor convertToJson function and move test file to separate directory
