+++ src/configLoader.ts
@@ -28,10 +28,10 @@ export const findConfigFiles = (path = './'): Promise<ConfigFileInterface> => {
   );
 };
 
-// export const yaml2Json = (inputFilePath: string) => {
-//   const pathSplits = inputFilePath.split(sep);
-//   const fileName = pathSplits[pathSplits.length - 1].replace('.yaml', '');
-//   return {
-//     [fileName]: yaml.load(fs.readFileSync(inputFilePath, { encoding: 'utf-8' }))
-//   };
-// };
+export const yaml2Json = (inputFilePath: string) => {
+  const pathSplits = inputFilePath.split(sep);
+  const fileName = pathSplits[pathSplits.length - 1].replace('.yaml', '');
+  return {
+    [fileName]: yaml.load(fs.readFileSync(inputFilePath, { encoding: 'utf-8' }))
+  };
+};
