+++ src/toJson.ts
@@ -2,7 +2,6 @@ import ErrnoException = NodeJS.ErrnoException;
 const toml = require('toml');
 const fs = require('fs');
 import yaml from 'js-yaml';
-import { sep } from 'path';
 
 export const convertToJson = async (
   path: string,
@@ -16,14 +15,11 @@ export const convertToJson = async (
         resolve({ [fileName]: JSON.parse(JSON.stringify(toml.parse(data))) });
       else if (fileExtension === '.json')
         resolve({ [fileName]: JSON.parse(data) });
+      else {
+        resolve({
+          [fileName]: yaml.load(fs.readFileSync(path, { encoding: 'utf-8' }))
+        });
+      }
     });
   });
-};
-
-export const yaml2Json = (inputFilePath: string) => {
-  const pathSplits = inputFilePath.split(sep);
-  const fileName = pathSplits[pathSplits.length - 1].replace('.yaml', '');
-  return {
-    [fileName]: yaml.load(fs.readFileSync(inputFilePath, { encoding: 'utf-8' }))
-  };
 };
